// Desarrollo_Web_en_entorno_cliente/Casa/Trabajos/Unit3 - Ionic/ionic-foodscore/src/app/restaurants/interfaces/restaurant.ts
import { User } from "src/app/auth/interfaces/user.interface";



export interface Restaurant {
    id?: number;
    name: string;
    description: string;
    cuisine: string;
    daysOpen: string[];
    image: string;
    phone: string;
    creator?: User;
    mine?: boolean;
    distance?: number;
    commented?: boolean;
    stars?: number;
    address: string;
    lat: number;
    lng: number;
}